V5, bug fixes
-Bug fixes:
- Promotion bug refixed
- Fixed reference bug, in my sort moves function I previously passed depth and ply by value. This caused a bug that was only recreateable in release mode and could be fixed by adding in print statements. I think the bug was caused by passing a variable by value to function1 then passing that variable by value to function2 without using it in function1. I am assuming without compiler optimizations, the variable is not initialized in function1 and therefore comes out to a garbage value. Insane.
- Fixed mate bug, was able to recreate this at depth 1 in this position 8/3Q4/k7/1p2p3/p3P3/P1P4P/1P2BPP1/6K1 b - - 0 1. I figured out that if all moves led to a quiescent search where one side will always end in check. I would never take stand pat but I only ever take the evaluation of a position as stand pat. I have added a fix by just returning the evaluation in such cases in qSearch.

+++ TranspositionTable.cpp
@@ -67,11 +67,11 @@ void recordEntry(int depth, int val, int hashFlag, int bestMove) {
 	}*/
 	if (currAge > newEntry->age || newEntry->depth <= depth) {
 		newEntry->key = zobristKey;
-		newEntry->depth = depth;
-		newEntry->flag = hashFlag;
+		newEntry->depth = (uint8_t)depth;
+		newEntry->flag = (uint8_t)hashFlag;
 		newEntry->value = val;
 		newEntry->bestMove = bestMove;
-		newEntry->age = currAge;
+		newEntry->age = (uint8_t)currAge;
 	}
 }
